+++ content/Dwt/Dwt.Api/Bootstrap/P1000ControllersBootstrapper.cs
@@ -1,0 +1,32 @@
+﻿using Dwt.Api.Controllers;
+using Microsoft.AspNetCore.Mvc;
+
+namespace Dwt.Api.Bootstrap;
+
+[Bootstrapper]
+public class ControllersBootstrapper
+{
+	public static void ConfigureBuilder(WebApplicationBuilder appBuilder)
+	{
+		appBuilder.Services.AddControllers()
+			.ConfigureApiBehaviorOptions(options =>
+			{
+				// configure custom response for invalid model state (usually input validation failed)
+				options.InvalidModelStateResponseFactory = context =>
+				{
+					var errors = context.ModelState.Values.SelectMany(v => v.Errors.Select(e => e.ErrorMessage));
+					return new BadRequestObjectResult(new ApiResp<object>
+					{
+						Status = 400,
+						Message = "Bad request.",
+						Data = errors
+					});
+				};
+			});
+	}
+
+	public static void DecorateApp(WebApplication app)
+	{
+		app.MapControllers();
+	}
+}