+++ content/Dwt/Dwt.Shared.EF/Commons.cs
@@ -1,0 +1,6 @@
+﻿namespace Dwt.Shared.EF;
+
+public sealed class Commons
+{
+	public const string TABLE_PREFIX = "dwt_";
+}