+++ content/Dwt/Dwt.Shared/Identity/DwtUser.cs
@@ -1,0 +1,7 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace Dwt.Shared.Identity;
+
+public class DwtUser : IdentityUser
+{
+}